requirements.txt
Django==3.0.7
django-solo==1.1.3
graphene-django==2.11.0
django-webpack-loader==0.7.0
pytz==2020.1